+++ include/menus/paged_list_menu.h
@@ -22,9 +22,9 @@ public:
   void Init();
   void Draw() override;
   void HandleInput(u32 keys) override;
-  u8 GetCurrentPage() const override;
-  u8 GetPageCount() const override;
-  void SelectItem(u8 index) override;
+  u16 GetCurrentPage() const override;
+  u16 GetPageCount() const override;
+  void SelectItem(u16 index) override;
 
   inline bool IsDirty() const { return dirty; }
   inline void SetDirty(bool d = true) { dirty = d; }
@@ -39,11 +39,11 @@ protected:
                             std::vector<u16> &pages) = 0;
   virtual bool ResolveTargetPage(u8 index, u16 *page_out);
 
-  u8 GetPageStart(u8 page_index) const;
-  u8 GetPageSize(u8 page_index) const;
-  u8 GetPageForIndex(u8 index) const;
+  u16 GetPageStart(u16 page_index) const;
+  u16 GetPageSize(u16 page_index) const;
+  u16 GetPageForIndex(u16 index) const;
   void UpdatePageSize();
-  std::vector<u8> page_sizes;
+  std::vector<u16> page_sizes;
 
 private:
   void ActivateSelected();
